% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r, R/temp2.r
\name{find_min_sites_analytical}
\alias{find_min_sites_analytical}
\title{find_min_sites_analytical}
\usage{
find_min_sites_analytical(
  S_grid,
  delta_target,
  sd_diff,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test")
)

find_min_sites_analytical(
  S_grid,
  delta_target,
  sd_diff,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test")
)
}
\arguments{
\item{delta_target}{Target mean change}

\item{sd_diff}{Standard deviation of difference in means}

\item{distribution}{Distribution for simulated data. Must be "normal" here.}

\item{useTest}{Which test to use. Must be "paired-t" here.}

\item{n_grid}{Number of sites grid}
}
\value{
Power for each number of sites in n_grid

Power for each number of sites in n_grid
}
\description{
Function to use power.t.test to calculate power for a paired t-test on the means across

Function to use power.t.test to calculate power for a paired t-test on the means across
}
