% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{getSD_difference}
\alias{getSD_difference}
\title{getSD_difference}
\usage{
getSD_difference(sd_within, nA, nB, sd_delta)
}
\arguments{
\item{sd_within}{standard deviation within sites}

\item{nA}{number of after samples per site}

\item{nB}{number of before samples per site}

\item{sd_delta}{standard deviation of true changes among sites}
}
\value{
Standard deviation of the difference in means
}
\description{
Estimate standard deviation of difference in means before and after change for
paired t-test of difference in means.
}
