% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{getSD_within}
\alias{getSD_within}
\title{getSD_within}
\usage{
getSD_within(baseline, siteVar = "site", responseVar = "y")
}
\arguments{
\item{baseline}{Data frame containing baseline measurements}

\item{siteVar}{Name of the site variable in baseline data}

\item{responseVar}{Name of the response variable in baseline data}
}
\value{
Estimated within-site standard deviation
}
\description{
Estimate within-site standard deviation from baseline data.
}
