% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{power_for_n_after_2samp}
\alias{power_for_n_after_2samp}
\title{power_for_n_after_2samp}
\usage{
power_for_n_after_2samp(
  S,
  nB,
  nA,
  delta,
  sd_pooled,
  sd_d = 0,
  alpha = 0.05,
  nsim = 2000,
  seed = 1
)
}
\arguments{
\item{S}{Number of sites}

\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{delta}{Hypothesized mean change}

\item{sd_pooled}{Standard deviation between samples}

\item{sd_d}{Between-site standard deviation of true changes (default 0)}

\item{alpha}{Significance level, defaults to 0.05}

\item{nsim}{Number of simulations to run}

\item{seed}{Random seed for reproducibility}
}
\value{
Estimated power (numeric scalar)
}
\description{
Estimate power for a two-sample (unpaired) t-test comparing all before vs. all
after measurements, ignoring possible within-site correlation.
}
