% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{power_for_nA}
\alias{power_for_nA}
\title{power_for_nA
Function to calculate power for given number of after samples for a paire t-test
of the difference in means before vs. after change using power.t.test}
\usage{
power_for_nA(nA, S, nB, delta, sd_within, sd_delta, alpha)
}
\arguments{
\item{nA}{Number of after samples per site}

\item{S}{Number of sites}

\item{nB}{Number of before samples per site}

\item{delta}{Hypothesized mean change}

\item{sd_within}{Standard deviation within sites}

\item{sd_delta}{Standard deviation of true changes among sites}

\item{alpha}{Significance level}
}
\value{
Power for the given number of after sample
}
\description{
power_for_nA
Function to calculate power for given number of after samples for a paire t-test
of the difference in means before vs. after change using power.t.test
}
