% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{power_for_percent_change}
\alias{power_for_percent_change}
\title{power_for_percent_change}
\usage{
power_for_percent_change(
  percent_change,
  baseline_mean = NULL,
  nA,
  S,
  nB,
  sd_within,
  sd_delta,
  alpha,
  typeTransform = c("none", "log", "asin", "sqrt"),
  addValue = 0
)
}
\arguments{
\item{percent_change}{Target percent change in the original response variable}

\item{baseline_mean}{Mean of the original response variable before the change}

\item{nA}{Number of after samples per site}

\item{S}{Number of sites}

\item{nB}{Number of before samples per site}

\item{sd_within}{Standard deviation within sites (in transformed scale)}

\item{sd_delta}{Standard deviation of true changes among sites}

\item{alpha}{Significance level}

\item{typeTransform}{Character indicating the transformation to apply to the response
variable, one of "none", "log","asin" or "sqrt" (default "none").}

\item{addValue}{Value to add to response variable before transforming to avoid issues}
}
\value{
Power for the given percent change
}
\description{
Function to calculate power for a given percent change in the mean
before vs. after change using power.t.test (paired t-test on site-level differences).
}
