% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{summarize_baseline}
\alias{summarize_baseline}
\title{summarize_baseline}
\usage{
summarize_baseline(
  baseline,
  siteVar = "site",
  responseVar = "y",
  groupVar = NULL,
  typeTransform = "none",
  addValue = 0
)
}
\arguments{
\item{baseline}{Data frame containing baseline measurements}

\item{siteVar}{Name of the site variable in baseline data}

\item{responseVar}{Name of the response variable in baseline data}

\item{groupVar}{Optional character vector of grouping variables to summarize by}

\item{addValue}{Value to add to response variable before transforming (default 0)}

\item{typeTansform}{Character indicating transformation the response variable, "none","log", "sqrt", or "asin" for asin(sqrt) (default "none")"}
}
\value{
A data frame containing the estimated mean and standard deviation of the
response variable across all sites and data points (grand_mean,grand_sd), and the
within and across site standard deviations (sd_within,
sd_between) in the original scale, and optionally the log or sqrt transformed
scale if typeTransform is "log" or "sqrt", with an added constant if
provide d(addValue). The number of sites in the baseline is provided (nB), or, if
the number of samples per site is not consistent, the number of complete
and the number of sites with fewer samples (nb_complete and nB_incomplete).
The proportion of positive values in the response variable is also calculated (prop_positive) to help assess whether a log transformation is appropriate.
If groupVar is provided, these statistics are calculated separately for each group defined by groupVar.
}
\description{
Summarize baseline data to estimate within-site and between-site standard deviations, and
optionally log-transform the response variable to estimate these on the log scale.
}
