% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_min_detectable_percent}
\alias{find_min_detectable_percent}
\title{find_min_detectable_percent}
\usage{
find_min_detectable_percent(
  S,
  nB,
  nA,
  sd_within,
  sd_delta = 0,
  sd_within_log,
  sd_delta_log = 0,
  logTransform = FALSE,
  logAdd = 0,
  target_power = 0.8,
  alpha = 0.05
)
}
\arguments{
\item{S}{Number of sites}

\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{sd_within}{Within-site standard deviation (log calculated if logTransform=TRUE)}

\item{sd_delta}{Between-site SD of true changes (default 0)}

\item{logTransform}{Logical indicating whether to calculate the detectable change on the log scale (default TRUE)}

\item{logAdd}{Value to add to response variable before log-transforming to avoid issues}

\item{target_power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
The minimum detectable percentage change (e.g., 30 for a 30\% change)
}
\description{
Calculates the smallest percentage change detectable for a given sample size.
}
