% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_min_detectable_percent_2samp}
\alias{find_min_detectable_percent_2samp}
\title{find_min_detectable_percent_2samp}
\usage{
find_min_detectable_percent_2samp(
  S,
  nB,
  nA,
  sd_pooled,
  baseline_mean,
  target_power = 0.8,
  alpha = 0.05,
  typeTransform = "arcsin"
)
}
\arguments{
\item{S}{Number of sites}

\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{sd_pooled}{Standard deviation among all data points}

\item{baseline_mean}{Mean of the response variable before the change (used to
convert the detectable absolute delta to a percent change)}

\item{target_power}{Desired power (default 0.8)}

\item{alpha}{Significance level (default 0.05)}
}
\value{
The minimum detectable percentage change (e.g., 30 for a 30\% change)
}
\description{
Calculates the smallest percentage change detectable for a given sample size
using a two-sample (unpaired) t-test that ignores within-site correlation.
}
