% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_min_sites_2samp}
\alias{find_min_sites_2samp}
\title{find_min_sites_2samp}
\usage{
find_min_sites_2samp(
  nB,
  nA,
  delta,
  sd_pooled,
  target_power = 0.8,
  alpha = 0.05,
  S_grid = 2:50
)
}
\arguments{
\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{delta}{Hypothesized mean change (absolute, on the scale of the response)}

\item{sd_pooled}{Standard deviation among all samples}

\item{target_power}{Desired power level (default 0.8)}

\item{alpha}{Significance level (default 0.05)}

\item{S_grid}{Grid of site numbers to evaluate (default 2:50)}
}
\value{
A list with \code{S_star} (minimum sites for target power) and \code{curve}
(data frame of S and power)
}
\description{
Find the minimum number of sites needed to achieve target power for a
two-sample (unpaired) t-test that pools all before and after measurements,
ignoring within-site correlation.
}
