% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_min_sites_2samp}
\alias{find_min_sites_2samp}
\title{find_min_sites_2samp}
\usage{
find_min_sites_2samp(
  nB,
  nA,
  S_before = NULL,
  delta,
  sd_pooled,
  target_power = 0.8,
  alpha = 0.05,
  S_grid = 2:50
)
}
\arguments{
\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{S_before}{Number of sites before, if you wish to keep this number
constant in the analysis. Defaults to NULL, which allows the number of sites
before and after to vary in the calculation. If S_before is provided,
the function will calculate power for a fixed number of sites before and varying number of sites after.}

\item{delta}{Hypothesized mean change (absolute, on the scale of the response)}

\item{sd_pooled}{Standard deviation among all samples}

\item{target_power}{Desired power level (default 0.8)}

\item{alpha}{Significance level (default 0.05)}

\item{S_grid}{Grid of site numbers to evaluate (default 2:50)}
}
\value{
A list with \code{S_star} (minimum sites for target power) and \code{curve}
(data frame of S and power). If nB and nA are the number of years before
after, then S_star is the minimum number of sites needed to achieve target
power with a sample size of  \code{n_before=nB*S} and \code{n_before=nA*S} in a 2-sample
t-test. If S_before is provided, then S_star is the minimum number of
sites needed to achieve target power with S_before sites before and
S_star sites after.
}
\description{
Find the minimum number of sites needed to achieve target power for a
two-sample (unpaired) t-test that pools all before and after measurements,
ignoring within-site correlation.
}
