% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_min_sites_analytical}
\alias{find_min_sites_analytical}
\title{find_min_sites_analytical}
\usage{
find_min_sites_analytical(S_grid, delta_target, sd_diff)
}
\arguments{
\item{delta_target}{Target mean change}

\item{sd_diff}{Standard deviation of difference in means}

\item{n_grid}{Number of sites grid}
}
\value{
Power for each number of sites in n_grid
}
\description{
Function to use power.t.test to calculate power for a paired t-test on the means across
}
