% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{find_n_after}
\alias{find_n_after}
\title{find_n_after}
\usage{
find_n_after(
  S,
  nB,
  delta,
  sd_w,
  sd_d = 0,
  target_power = 0.8,
  alpha = 0.05,
  n_grid = 1:50,
  nsim = 2000,
  seed = 1,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test"),
  nbinom_mu = NULL,
  nbinom_disp = NULL,
  binomial_size = NULL,
  binomial_prob = NULL
)
}
\arguments{
\item{S}{Number of sites}

\item{nB}{Number of before measurements per site}

\item{delta}{Hypothesized mean change}

\item{sd_w}{Within-site standard deviation}

\item{sd_d}{Between-site standard deviation of true changes}

\item{target_power}{Desired power level to achieve}

\item{alpha}{Significance level, defaults to 0.05}

\item{n_grid}{Grid of after measurements to evaluate}

\item{nsim}{Number of simulations to run}

\item{seed}{Random seed for reproducibility}

\item{distribution}{Distribution for simulated data. One of
"normal", "nbinom", or "binomial".}

\item{useTest}{Which test to use. One of "paired-t", "wilcoxon", or "prop.test".}

\item{nbinom_mu}{Mean parameter for negative binomial (mu)}

\item{nbinom_disp}{Dispersion (size) parameter for negative binomial}

\item{binomial_size}{Size parameter (trials) for binomial}

\item{binomial_prob}{Probability parameter for binomial}
}
\value{
A list containing the minimum number of after measurements needed to achieve the target power and a data frame of the power curve
}
\description{
Function to find the minimum number of after measurements needed to achieve target
power, for a paired t-test on the before and after means.
}
