% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r, R/temp2.r
\name{power_for_n_after}
\alias{power_for_n_after}
\title{power_for_n_after}
\usage{
power_for_n_after(
  S,
  nB,
  nA,
  delta,
  sd_w,
  sd_d = 0,
  alpha = 0.05,
  nsim = 2000,
  seed = 1,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test"),
  nbinom_mu = NULL,
  nbinom_disp = NULL,
  binomial_size = NULL,
  binomial_prob = NULL
)

power_for_n_after(
  S,
  nB,
  nA,
  delta,
  sd_w,
  sd_d = 0,
  alpha = 0.05,
  nsim = 2000,
  seed = 1,
  distribution = c("normal", "nbinom", "binomial"),
  useTest = c("paired-t", "wilcoxon", "prop.test"),
  nbinom_mu = NULL,
  nbinom_disp = NULL,
  binomial_size = NULL,
  binomial_prob = NULL
)
}
\arguments{
\item{S}{Number of sites}

\item{nB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{delta}{Hypothesized mean change}

\item{sd_w}{Within-site standard deviation}

\item{sd_d}{Between-site standard deviation of true changes}

\item{alpha}{Significance level, defaults to 0.05}

\item{nsim}{Number of simulations to run}

\item{seed}{Random seed for reproducibility}

\item{distribution}{Distribution for simulated data. One of
"normal", "nbinom", or "binomial".}

\item{useTest}{Which test to use. One of "paired-t", "wilcoxon", or "prop.test".}

\item{nbinom_mu}{Mean parameter for negative binomial (mu)}

\item{nbinom_disp}{Dispersion (size) parameter for negative binomial}

\item{binomial_size}{Size parameter (trials) for binomial}

\item{binomial_prob}{Probability parameter for binomial}
}
\value{
Estimated power for the given number of after measurements

Estimated power for the given number of after measurements
}
\description{
Estimate power for a paired t-test on the average of a given number of
measurements at a fixed number of sites before and after a changepoint.

Estimate power for a paired t-test on the average of a given number of
measurements at a fixed number of sites before and after a changepoint.
}
