% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PowerAfterFunctions.r
\name{power_for_sites}
\alias{power_for_sites}
\title{power_for_sites}
\usage{
power_for_sites(
  S,
  mB,
  nA,
  delta,
  sd_w,
  sd_d = 0,
  alpha = 0.05,
  nsim = 2000,
  seed = 1
)
}
\arguments{
\item{S}{Number of sites}

\item{mB}{Number of before measurements per site}

\item{nA}{Number of after measurements per site}

\item{delta}{Hypothesized mean change}

\item{sd_w}{Within-site standard deviation}

\item{sd_d}{Between-site standard deviation of true changes}

\item{alpha}{Significance level, defaults to 0.05}

\item{nsim}{Number of simulations to run}

\item{seed}{Random seed for reproducibility}
}
\value{
Estimated power for the given number of sites
}
\description{
power_for_sites
}
